+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Lizard test reporting in CI and Minitest reporter plugin in test_helper
+
 ### Removed
 
 - Ruby 3.2 support; minimum is now 3.3